Move host to the metadata package
